Move shadcn stories into stories
